+++ faeBeta/faeBeta/SelectLocationViewController.swift
@@ -29,7 +29,7 @@ class SelectLocationViewController: UIViewController {
     var currentLatitude: CLLocationDegrees = 34.0205378
     var currentLongitude: CLLocationDegrees = -118.2854081
     var currentLocation2D = CLLocationCoordinate2DMake(34.0205378, -118.2854081)
-    var zoomLevel: Float = 13.8
+    var zoomLevel: Double = 13.8
     var latitudeForPin: CLLocationDegrees = 0
     var longitudeForPin: CLLocationDegrees = 0
     var willAppearFirstLoad = false
